画像のURLチェックを修正し、特定の条件下での処理を改善しました
