+++ next/src/app/[locale]/_lib/html-to-mdast-with-segments.ts
@@ -7,13 +7,10 @@ import { removePosition } from "unist-util-remove-position";
 
 import { remarkHashAndSegments } from "@/app/[locale]/_lib/remark-hash-and-segments";
 import type { SegmentDraft } from "@/app/[locale]/_lib/remark-hash-and-segments";
-import { uploadImage } from "@/app/[locale]/_lib/upload";
+import { remarkAutoUploadImages } from "@/app/[locale]/_lib/remark-auto-upload-images";
 import type { Prisma } from "@prisma/client";
 import type { Root as MdastRoot } from "mdast";
 import pLimit from "p-limit";
-import sharp from "sharp";
-import type { Plugin } from "unified";
-import { visit } from "unist-util-visit";
 import { VFile } from "vfile";
 const limit = pLimit(5);
 interface Params {
@@ -67,36 +64,3 @@ export async function fileFromUrl(url: string): Promise<File> {
 }
 // S3 / R2 など
 
-export const remarkAutoUploadImages: Plugin<[]> = () => {
-	return async (tree: MdastRoot) => {
-		// ここに来た時点で必ず MDAST
-		const tasks: Promise<void>[] = [];
-
-		visit(tree, "image", (node) => {
-			if (!node.url.includes("evame/uploads")) return;
-			tasks.push(
-				limit(async () => {
-					const file = await fileFromUrl(node.url);
-
-					// 例：サーバで再エンコードして 2 MB 以下に
-					const buf = await sharp(await file.arrayBuffer())
-						.resize({ width: 2560, withoutEnlargement: true })
-						.jpeg({ quality: 80, mozjpeg: true })
-						.toBuffer();
-
-					const jpegName = file.name.replace(/\.[^.]+$/, ".jpg");
-					const compact = new File([buf], jpegName, { type: "image/jpeg" });
-
-					const result = await uploadImage(compact);
-					if (result.success) {
-						node.url = result.data.imageUrl;
-					} else {
-						console.error(result.message);
-					}
-				}),
-			);
-		});
-
-		await Promise.all(tasks);
-	};
-};
