+++ next/src/app/[locale]/_lib/remark-auto-upload-images.ts
@@ -1,0 +1,42 @@
+import sharp from "sharp";
+import type { Plugin } from "unified";
+import { visit } from "unist-util-visit";
+import { uploadImage } from "./upload";
+import { fileFromUrl } from "./html-to-mdast-with-segments";
+import pLimit from "p-limit";
+import type { Root as MdastRoot } from "mdast";
+const limit = pLimit(5);
+
+export const remarkAutoUploadImages: Plugin<[]> = () => {
+	return async (tree: MdastRoot) => {
+		// ここに来た時点で必ず MDAST
+		const tasks: Promise<void>[] = [];
+
+		visit(tree, "image", (node) => {
+			if (node.url.includes("evame/uploads")) return;
+			tasks.push(
+				limit(async () => {
+					const file = await fileFromUrl(node.url);
+
+					// 例：サーバで再エンコードして 2 MB 以下に
+					const buf = await sharp(await file.arrayBuffer())
+						.resize({ width: 2560, withoutEnlargement: true })
+						.jpeg({ quality: 80, mozjpeg: true })
+						.toBuffer();
+
+					const jpegName = file.name.replace(/\.[^.]+$/, ".jpg");
+					const compact = new File([buf], jpegName, { type: "image/jpeg" });
+
+					const result = await uploadImage(compact);
+					if (result.success) {
+						node.url = result.data.imageUrl;
+					} else {
+						console.error(result.message);
+					}
+				}),
+			);
+		});
+
+		await Promise.all(tasks);
+	};
+};
